Shifted to random stat generation
Characters now come with stats in a random range from 3 to 6 by default.  Might want to add an option to create a specific stat spread for a character.

Roster implemented, contains all current characters.  This could be done with a simple list right now, but this should be able to give future functionality.