Reorganized Data, updated Readme
Data files now all in the proper folder.

Updated the ReadMe file to give an overview of the project and what's implemented so far.

Added Team (smaller version of Roster for each game)

Added connections for Talent Trees.  This is currently just represented by a list of ints, not sure if it will be required in the future to pass Talents themselves.